Movr Auth check functions around
